Fix folder name + remove empty file
