+++ src/modules/accounting/accounting.py
@@ -10,7 +10,7 @@ from src.services.validator_state import LidoValidatorStateService
 from src.modules.submodules.consensus import ConsensusModule
 from src.modules.submodules.oracle_module import BaseModule
 from src.services.withdrawal import Withdrawal
-from src.modules.accounting.bunker import BunkerService
+from services.bunker import BunkerService
 from src.typings import BlockStamp, Gwei, ReferenceBlockStamp
 from src.utils.abi import named_tuple_to_dataclass
 from src.web3py.typings import Web3
